+++ src/components/SliderEditor.tsx
@@ -2,17 +2,17 @@ import React from 'react';
 import { StandardEditorProps } from '@grafana/data';
 import { Slider } from '@grafana/ui';
 
-export const SliderEditor: React.FC<StandardEditorProps<any, any>> = props => {
+export const SliderEditor: React.FC<StandardEditorProps> = (props) => {
   return (
     <div style={{ marginRight: 15 }}>
       <Slider
         tooltipAlwaysVisible={false}
         min={0}
         max={100}
-        onAfterChange={values => {
-          props.onChange(values[0]);
+        onAfterChange={(values) => {
+          props.onChange(values);
         }}
-        value={[props?.value]}
+        value={props?.value}
       />
     </div>
   );