+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## [v1.0.1](https://github.com/NubeIO/grafana-flow-framework-read-write-panel/tree/v1.0.1) (2022-01-09)
+
+- Make multi-switch text style as user expects
+- Fix: slider issue for grafana upgrade v7.5.x
+
 ## [v1.0.0](https://github.com/NubeIO/grafana-flow-framework-read-write-panel/tree/v1.0.0) (2022-01-08)
 
 - First initial release